requirement.txt
Flask-RESTful==0.3.5
flask-orator==0.2.0
Flask-JWT-Extended==3.3.4
mysqlclient==1.3.13
var-dump==1.1
Cerberus==1.1
redis==2.10.5
logstash-formatter==0.5.16